:recycle: PairRoomOnboarding 도메인 로직 usePairRoomInformation 커스텀 훅으로 분리
