+++ frontend/src/components/PairRoomOnboarding/ProgressBar/ProgressBar.stories.tsx
@@ -1,7 +1,5 @@
 import type { Meta, StoryObj } from '@storybook/react';
 
-import type { Step } from '@/pages/PairRoomOnboarding/PairRoomOnboarding.type';
-
 import ProgressBar from './ProgressBar';
 
 const meta = {
@@ -15,6 +13,6 @@ type Story = StoryObj<typeof ProgressBar>;
 
 export const Default: Story = {
   args: {
-    step: 'ROLE' as Step,
+    step: 'ROLE',
   },
 };
