+++ frontend/src/constants/PairRoomOnboarding/step.ts
@@ -1,6 +1,4 @@
-import { StepOption } from '@/pages/PairRoomOnboarding/PairRoomOnboarding.type';
-
-export const OPTIONS: StepOption[] = [
+export const OPTIONS = [
   {
     id: 'ROLE',
     label: '역할 설정',
