+++ frontend/src/pages/PairRoomOnboarding/PairRoomOnboarding.tsx
@@ -1,14 +1,12 @@
 import { useState } from 'react';
 
-import { Role } from '@/pages/PairRoomOnboarding/PairRoomOnboarding.type';
-
 import Button from '@/components/common/Button/Button';
 import PairNameInput from '@/components/PairRoomOnboarding/PairNameInput/PairNameInput';
 import PairRoleInput from '@/components/PairRoomOnboarding/PairRoleInput/PairRoleInput';
 import TimerDurationInput from '@/components/PairRoomOnboarding/TimerDurationInput/TimerDurationInput';
 
-import usePairNameInputs from '@/hooks/Main/usePairNameInputs';
 import useAutoMoveIndex from '@/hooks/PairRoomOnboarding/useAutoMoveIndex';
+import usePairRoomInformation from '@/hooks/PairRoomOnboarding/usePairRoomInformation';
 
 import { BUTTON_TEXT } from '@/constants/button';
 
@@ -16,47 +14,25 @@ import * as S from './PairRoomOnboarding.styles';
 
 const PairRoomOnboarding = () => {
   const [isTyping, setIsTyping] = useState(false);
-  const [driver, setDriver] = useState('');
-  const [navigator, setNavigator] = useState('');
-  const [timerDuration, setTimerDuration] = useState('');
 
-  const { firstPairName, secondPairName, isPairNameValid, handleFirstPairName, handleSecondPairName } =
-    usePairNameInputs();
+  const {
+    firstPairName,
+    secondPairName,
+    driver,
+    navigator,
+    timerDuration,
+    isPairNameValid,
+    isPairRoleValid,
+    isTimerDurationValid,
+    handleFirstPairName,
+    handleSecondPairName,
+    handlePairRole,
+    handleTimerDuration,
+  } = usePairRoomInformation();
 
-  const validationList = [isPairNameValid, driver !== '' && navigator !== '', timerDuration !== ''];
+  const validationList = [isPairNameValid, isPairRoleValid, isTimerDurationValid];
+
   const { moveIndex } = useAutoMoveIndex(0, validationList, isTyping);
-
-  const handleFirstPair = (event: React.ChangeEvent<HTMLInputElement>) => {
-    if (firstPairName.value === driver || firstPairName.value === navigator) {
-      setDriver('');
-      setNavigator('');
-    }
-
-    handleFirstPairName(event);
-  };
-
-  const handleSecondPair = (event: React.ChangeEvent<HTMLInputElement>) => {
-    if (secondPairName.value === driver || secondPairName.value === navigator) {
-      setDriver('');
-      setNavigator('');
-    }
-
-    handleSecondPairName(event);
-  };
-
-  const handleRole = (pairName: string, role: Role) => {
-    const otherPair = firstPairName.value === pairName ? secondPairName.value : firstPairName.value;
-
-    if (role === 'DRIVER') {
-      setDriver(pairName);
-      setNavigator(otherPair);
-    } else {
-      setDriver(otherPair);
-      setNavigator(pairName);
-    }
-  };
-
-  const handleTimerDuration = (timerDuration: string) => setTimerDuration(timerDuration);
 
   return (
     <S.Layout>
@@ -66,8 +42,8 @@ const PairRoomOnboarding = () => {
           <PairNameInput
             firstPairName={firstPairName}
             secondPairName={secondPairName}
-            onFirstPair={handleFirstPair}
-            onSecondPair={handleSecondPair}
+            onFirstPair={handleFirstPairName}
+            onSecondPair={handleSecondPairName}
             onFocus={() => setIsTyping(true)}
             onBlur={() => setIsTyping(false)}
           />
@@ -77,7 +53,7 @@ const PairRoomOnboarding = () => {
               secondPair={secondPairName.value}
               driver={driver}
               navigator={navigator}
-              onRole={handleRole}
+              onRole={handlePairRole}
             />
           )}
           {moveIndex >= 2 && <TimerDurationInput timerDuration={timerDuration} onTimerDuration={handleTimerDuration} />}
