--- frontend/src/pages/PairRoomOnboarding/PairRoomOnboarding.type.ts
@@ -1,8 +1,0 @@
-export type Step = 'MISSION' | 'ROLE' | 'TIMER';
-
-export interface StepOption {
-  id: Step;
-  label: string;
-}
-
-export type Role = 'DRIVER' | 'NAVIGATOR';
