--- frontend/src/utils/PairRoomOnboarding/validate.ts
@@ -1,6 +1,0 @@
-export const validateTime = (time: string) => {
-  if (Number.isNaN(time)) return false;
-  if (Number(time) <= 0) return false;
-
-  return true;
-};
