+++ frontend/src/validations/validateTimerDuration.ts
@@ -1,0 +1,6 @@
+export const validateTimerDuration = (timerDuration: string) => {
+  if (Number.isNaN(timerDuration)) return false;
+  if (Number(timerDuration) <= 0) return false;
+
+  return true;
+};
